+++ RandomSkunk.Results/Result.cs
@@ -119,23 +119,13 @@ public partial struct Result : IEquatable<Result>
     /// <param name="errorMessage">The optional error message.</param>
     /// <param name="errorCode">The optional error code.</param>
     /// <param name="errorIdentifier">The optional identifier of the error.</param>
-    /// <param name="errorType">
-    /// The optional type of the error. If <see langword="null"/>, then the
-    /// <see cref="MemberInfo.Name"/> of the <see cref="Type"/> of the current instance
-    /// is used instead.
-    /// </param>
-    /// <param name="innerError">
-    /// The optional error that is the cause of the current error.
-    /// </param>
     /// <returns>A <c>Fail</c> result.</returns>
     public static Result Fail(
         Exception exception,
         string? errorMessage = null,
         int? errorCode = null,
-        string? errorIdentifier = null,
-        string? errorType = null,
-        Error? innerError = null) =>
-        Fail(FromException(exception, errorMessage, errorCode, errorIdentifier, errorType, innerError));
+        string? errorIdentifier = null) =>
+        Fail(FromException(exception, errorMessage, errorCode, errorIdentifier));
 
     /// <summary>
     /// Creates a <c>Fail</c> result.